adjusted RI test stacking
