+++ 03_myopic-multi_market.py
@@ -17,11 +17,11 @@ import os
 import pandas as pd
 from loguru import logger
 
-from src.coordinated_multi_market.rolling_intrinsic.testing_rolling_intrinsic_h_intelligent_stacking import (
-    simulate_period_stacked_hourly_products,
-)
-from src.coordinated_multi_market.rolling_intrinsic.testing_rolling_intrinsic_qh_intelligent_stacking import (
-    simulate_period_stacked_quarterhourly_products,
+#from src.coordinated_multi_market.rolling_intrinsic.testing_rolling_intrinsic_h_intelligent_stacking import (
+#    simulate_period_stacked_hourly_products,
+#)
+from src.coordinated_multi_market.rolling_intrinsic.new_testing_rolling_intrinsic_qh_intelligent_stacking import (
+    simulate_stacked_quarterhourly_products,
 )
 from src.shared.config import (
     BUCKET_SIZE,
@@ -49,13 +49,13 @@ if __name__ == "__main__":
     )
 
     # Run the rolling intrinsic simulation with intelligent stacking (QH products)
-    simulate_period_stacked_quarterhourly_products(
+    simulate_stacked_quarterhourly_products(
         da_bids_path=INPUT_DIR_DA,
         output_path=ri_qh_output_path,
         start_day=START,
         end_day=END,
-        threshold=0,
-        threshold_abs_min=0,
+        #threshold=0,
+        #threshold_abs_min=0,
         discount_rate=0,
         bucket_size=BUCKET_SIZE,
         c_rate=C_RATE,
