+++ src/coordinated_multi_market/rolling_intrinsic/new_training_rolling_intrinsic_qh_intelligent_stacking.py
@@ -1,88 +1,96 @@
 import os
-import warnings
-from typing import Optional
 
 import numpy as np
 import pandas as pd
-import psycopg2
-from dotenv import load_dotenv
 from loguru import logger
 from pulp import GUROBI, PULP_CBC_CMD, LpMaximize, LpProblem, LpVariable, lpSum
 
-from src.shared.config import (
-    COORDINATED_STACKED_RI_QH_TRAINING_OUTPUT_CSV,
-    LOGGING_PATH_COORDINATED,
-)
-from src.shared.folder_versioning import get_current_dir_version
+import psycopg2
+from dotenv import load_dotenv
+from sqlalchemy import create_engine
+import time
+
+from typing import Optional
+
 
 load_dotenv()
 
-
-warnings.simplefilter(action="ignore", category=FutureWarning)
-
-PASSWORD = os.getenv("SQL_PASSWORD")
-if PASSWORD:
-    password_for_url = f":{PASSWORD}"
-else:
-    password_for_url = ""
-
-POSTGRES_DB_NAME = os.getenv("POSTGRES_DB_NAME")
-CONNECTION = f"postgres://elli{password_for_url}@127.0.0.1/{POSTGRES_DB_NAME}"
-CONNECTION_ALCHEMY = f"postgresql://elli{password_for_url}@127.0.0.1/{POSTGRES_DB_NAME}"
+PRECOMPUTED_VWAP_PATH = os.path.join("data", "precomputed_vwaps")
 
 
-def get_average_prices(
-    cursor, side, execution_time_start, execution_time_end, end_date, min_trades=10
-):
-    # set start_of_day to end_date minus 1 day
-    start_of_day = pd.to_datetime(end_date) - pd.Timedelta(hours=2)
 
-    # set hour and minute to 0 (europe/berlin time)
-    start_of_day = start_of_day.replace(hour=0, minute=0)
 
-    end_of_day = start_of_day
-
-    end_of_day = end_of_day.replace(hour=23, minute=45)
-
-    year = start_of_day.year
-
-    table_name = "transactions_intraday_de"
-
-    # transform dates to work with str format
-    execution_time_start_str = execution_time_start.strftime("%Y-%m-%d %H:%M:%S")
-    execution_time_end_str = execution_time_end.strftime("%Y-%m-%d %H:%M:%S")
-    start_of_day_str = start_of_day.strftime("%Y-%m-%d %H:%M:%S")
-    end_date_str = end_date.strftime("%Y-%m-%d %H:%M:%S")
-
-    cursor.execute(
-        f"""
-        SELECT
-            deliverystart,
-            SUM(weighted_avg_price * volume) / SUM(volume) AS weighted_avg_price
-        FROM
-            {table_name}
-        WHERE
-            (executiontime BETWEEN '{execution_time_start_str}' AND '{execution_time_end_str}')
-            AND side = '{side}'
-            AND deliverystart >= '{start_of_day_str}'
-            AND deliverystart < '{end_date_str}'
-        GROUP BY
-            deliverystart
-        HAVING
-            SUM(trade_count) >= {min_trades};
+def load_vwap_matrix_for_day(current_day, base_path=PRECOMPUTED_VWAP_PATH):
     """
+    Lädt die vorcomputierte VWAP-Matrix für einen Tag aus einer Parquet-Datei.
+
+    Parquet-Format:
+        - index: bucket_end (Execution-Fenster-Ende) als String
+        - columns: deliverystart (Produkt) als String
+        - Werte: VWAP (float / NaN)
+    """
+    fname = os.path.join(base_path, f"vwaps_{current_day:%Y-%m-%d}.parquet")
+
+    if not os.path.exists(fname):
+        raise FileNotFoundError(
+            f"VWAP-Parquet für {current_day:%Y-%m-%d} nicht gefunden: {fname}"
+        )
+
+    matrix = pd.read_parquet(fname)
+
+    # Index / Columns wieder zu Timestamps machen (immer über UTC gehen)
+    # Strings wie "2021-03-27 00:15:00+01:00" -> UTC -> Europe/Berlin
+    matrix.index = (
+        pd.to_datetime(matrix.index, utc=True)
+        .tz_convert("Europe/Berlin")
     )
-    result = cursor.fetchall()
+    matrix.columns = (
+        pd.to_datetime(matrix.columns, utc=True)
+        .tz_convert("Europe/Berlin")
+    )
 
-    df = pd.DataFrame(result, columns=["product", "price"])
+    return matrix
 
-    # set index to product
-    df.set_index("product", inplace=True)
 
-    # set index to be all 15 minute intervals from start_of_day to end_of_day, filling missing values with NaN
-    df = df.reindex(pd.date_range(start_of_day, end_of_day, freq="60min"))
 
-    return df
+
+
+
+def get_vwap_from_matrix_for_bucket(vwap_matrix, execution_time_end, end_date):
+    """
+    Holt aus der VWAP-Matrix die passende Zeile für ein bestimmtes execution_time_end
+    und formt sie in dasselbe Format wie get_average_prices(...):
+
+        DataFrame:
+            index: 15-Minuten-Intervalle von start_of_day bis end_of_day
+            Spalte: "price"
+
+    end_date ist wie im Original: das Ende des Betrachtungszeitraums (trading_end),
+    wird verwendet, um start_of_day / end_of_day zu bestimmen.
+    """
+
+    # start_of_day / end_of_day genauso wie in get_average_prices
+    start_of_day = pd.to_datetime(end_date) - pd.Timedelta(hours=2)
+    start_of_day = start_of_day.replace(hour=0, minute=0)
+    end_of_day = start_of_day.replace(hour=23, minute=45)
+
+    # Falls das Execution-Fenster nicht in der Matrix existiert:
+    if execution_time_end not in vwap_matrix.index:
+        idx = pd.date_range(start_of_day, end_of_day, freq="15min")
+        return pd.DataFrame(index=idx, columns=["price"], data=np.nan)
+
+    # Zeile für dieses execution_time_end holen
+    row = vwap_matrix.loc[execution_time_end]       # Series: index = deliverystart, values = vwap
+
+    # In DataFrame mit Spalte "price" verwandeln
+    vwap_df = row.to_frame(name="price")
+
+    # Sicherstellen, dass wir das vollständige 15-Minuten-Raster haben
+    full_index = pd.date_range(start_of_day, end_of_day, freq="15min")
+    vwap_df = vwap_df.reindex(full_index)
+
+    return vwap_df
+
 
 
 def calculate_discounted_price(price, current_time, delivery_time, discount_rate):
@@ -101,6 +109,65 @@ def calculate_discounted_price(price, current_time, delivery_time, discount_rate
     return price * discount_factor
 
 
+
+
+def derive_day_ahead_trades_from_drl_output(
+    output: pd.DataFrame, current_day: pd.Timestamp
+) -> pd.DataFrame:
+    """
+    Leitet aus dem DRL-Output die Day-Ahead-Trades für current_day ab
+    und gibt sie im gleichen Format zurück wie Intraday-Trades:
+    Spalten: execution_time, side, quantity, price, product, profit
+    """
+    day_ahead_trades = {}
+
+    # Zeile(n) des DRL-Outputs für genau diesen Tag holen
+    # Erwartung: Index enthält current_day.date().isoformat() als Label
+    df = output.loc[current_day.date().isoformat()].copy().round(2)
+
+    # Nur Einträge mit tatsächlichem Trade
+    mask = df.capacity_trade != 0
+    df = df[mask]
+
+    # buy/sell + Betrag
+    df["side"] = ["buy" if x < 0 else "sell" for x in df.capacity_trade]
+    df["net_volume"] = [abs(x) for x in df.capacity_trade]
+
+    # Stundenprofit
+    df["profit"] = df.capacity_trade * df.epex_spot_60min_de_lu_eur_per_mwh
+
+    # time-Spalte in normaler Form
+    df.reset_index(inplace=True)
+
+    # Day-Ahead-Clearing: Vortag um 13:00
+    day_ahead_market_clearing = (current_day - pd.Timedelta(days=1)).replace(hour=13)
+
+    for _, row in df.iterrows():
+        # Stundenposition auf 4 Viertelstunden verteilen
+        product_indexes = pd.date_range(row["time"], periods=4, freq="15min")
+        for product_index in product_indexes:
+            day_ahead_trades.update(
+                {
+                    product_index: {
+                        "execution_time": day_ahead_market_clearing,
+                        "side": row["side"],
+                        "quantity": row["net_volume"],
+                        "price": row["epex_spot_60min_de_lu_eur_per_mwh"],
+                        "product": product_index,
+                        "profit": row["profit"] / 4,
+                    }
+                }
+            )
+
+    return pd.DataFrame(day_ahead_trades).T.reset_index(drop=True)
+
+
+
+
+
+
+
+
 def run_optimization_quarterhours_repositioning(
     prices_qh,
     execution_time,
@@ -108,8 +175,6 @@ def run_optimization_quarterhours_repositioning(
     c_rate,
     roundtrip_eff,
     max_cycles,
-    threshold,
-    threshold_abs_min,
     discount_rate,
     prev_net_trades=pd.DataFrame(
         columns=["sum_buy", "sum_sell", "net_buy", "net_sell", "product"]
@@ -171,6 +236,8 @@ def run_optimization_quarterhours_repositioning(
 
     e = 0.01
 
+    efficiency = roundtrip_eff**0.5
+
     # Objective function
     # Adjusted objective component for cases where previous trades < e
     adjusted_obj = [
@@ -179,29 +246,24 @@ def run_optimization_quarterhours_repositioning(
                 current_sell_qh[i]
                 * (
                     prices_qh_adj.loc[i, "price"]
-                    - max(
-                        abs((threshold / 100) * abs(prices_qh.loc[i, "price"])),
-                        threshold_abs_min,
-                    )
-                    / 2
+                    - 0.1 / 2  # assumed transaction costs per trade
                     - e
                 )
+                # * efficiency  # monetary efficiency consideration
             )
             - (
                 current_buy_qh[i]
                 * (
                     prices_qh_adj_buy.loc[i, "price"]
-                    + max(
-                        abs((threshold / 100) * abs(prices_qh.loc[i, "price"])),
-                        threshold_abs_min,
-                    )
-                    / 2
+                    + 0.1 / 2  # assumed transaction costs per trade
                     + e
                 )
+                
+                #/ efficiency  # monetary efficiency consideration
             )
         )
         * 1.0
-        / 1.0
+        / 4.0
         for i in prices_qh.index
         if not pd.isna(prices_qh.loc[i, "price"])
         and (
@@ -214,10 +276,10 @@ def run_optimization_quarterhours_repositioning(
     original_obj = [
         (
             current_sell_qh[i] * (prices_qh.loc[i, "price"] - e)
-            - current_buy_qh[i] * prices_qh.loc[i, "price"]
+            - current_buy_qh[i] * (prices_qh.loc[i, "price"] + e)
         )
         * 1.0
-        / 1.0
+        / 4.0
         for i in prices_qh.index
         if not pd.isna(prices_qh.loc[i, "price"])
         and (
@@ -232,14 +294,12 @@ def run_optimization_quarterhours_repositioning(
     # Constraints
     previous_index = prices_qh.index[0]
 
-    efficiency = roundtrip_eff**0.5
-
     for i in prices_qh.index[1:]:
         m_battery += (
             battery_soc[i]
             == battery_soc[previous_index]
-            + net_buy[previous_index] * efficiency * 1.0 / 1.0
-            - net_sell[previous_index] * 1.0 / 1.0 / efficiency,
+            + net_buy[previous_index] * efficiency *1.0 / 4.0
+            - net_sell[previous_index] * 1.0 / 4.0 / efficiency,
             f"BatteryBalance_{i}",
         )
         previous_index = i
@@ -256,7 +316,7 @@ def run_optimization_quarterhours_repositioning(
             m_battery += net_buy[i] <= cap * c_rate, f"BuyRate_{i}"
             m_battery += net_sell[i] <= cap * c_rate, f"SellRate_{i}"
             m_battery += (
-                net_sell[i] * 1.0 / efficiency / 1.0 <= battery_soc[i],
+                net_sell[i] * 1.0 / efficiency  / 4.0 <= battery_soc[i],
                 f"SellVsSOC_{i}",
             )
 
@@ -285,16 +345,15 @@ def run_optimization_quarterhours_repositioning(
 
     # set efficiency as sqrt of roundtrip efficiency
     m_battery += (
-        lpSum(net_buy[i] * efficiency * 1.0 / 1.0 for i in prices_qh.index)
-        <= max_cycles * cap,
+        lpSum(net_buy[i] * efficiency * 1.0 / 4.0 for i in prices_qh.index) <= max_cycles * cap,
         "MaxCycles",
     )
 
     # Solve the problem
-    # m_battery.solve(GUROBI(msg=0))
+    m_battery.solve(GUROBI(msg=0))
 
     # Solve the problem
-    m_battery.solve(PULP_CBC_CMD(msg=0))
+    #m_battery.solve(PULP_CBC_CMD(msg=0))
 
     # print(f"Status: {LpStatus[m_battery.status]}")
     # print(f"Objective value: {m_battery.objective.value()}")
@@ -317,7 +376,7 @@ def run_optimization_quarterhours_repositioning(
                 "quantity": [current_buy_qh[i].value()],
                 "price": [prices_qh.loc[i, "price"]],
                 "product": [i],
-                "profit": [-current_buy_qh[i].value() * prices_qh.loc[i, "price"] / 1],
+                "profit": [-current_buy_qh[i].value() * prices_qh.loc[i, "price"] / 4],
             }
 
             # append new trade using concat
@@ -331,7 +390,7 @@ def run_optimization_quarterhours_repositioning(
                 "quantity": [current_sell_qh[i].value()],
                 "price": [prices_qh.loc[i, "price"]],
                 "product": [i],
-                "profit": [current_sell_qh[i].value() * prices_qh.loc[i, "price"] / 1],
+                "profit": [current_sell_qh[i].value() * prices_qh.loc[i, "price"] / 4],
             }
 
             # append new trade using concat
@@ -391,7 +450,7 @@ def get_net_trades(trades, end_date):
     end_of_day = end_of_day.replace(hour=23, minute=45)
 
     net_trades = net_trades.reindex(
-        pd.date_range(start_of_day, end_of_day, freq="60min")
+        pd.date_range(start_of_day, end_of_day, freq="15min")
     )
 
     # fill NaN values with 0
@@ -404,28 +463,23 @@ def get_net_trades(trades, end_date):
     return net_trades
 
 
-def simulate_period_hourly_products(
+def simulate_days_stacked_quarterhourly_products(
     start_day,
     end_day,
-    threshold,
-    threshold_abs_min,
     discount_rate,
     bucket_size,
     c_rate,
     roundtrip_eff,
     max_cycles,
     min_trades,
-    day_ahead_trades_drl: Optional[pd.DataFrame] = None,
+    precomputed_vwap_path=PRECOMPUTED_VWAP_PATH,
+    drl_output: Optional[pd.DataFrame] = None,
 ):
-    conn = psycopg2.connect(CONNECTION)
-    cursor = conn.cursor()
-    cursor.execute("ROLLBACK")
+   
     log_message = (
-        "Running Rolling intrinsic H with the following parameters:\n"
+        "Running Rolling intrinsic QH with the following parameters:\n"
         "Start Day: {start_day}\n"
         "End Day: {end_day}\n"
-        "Threshold: {threshold}\n"
-        "Threshold Absolute Minimum: {threshold_abs_min}\n"
         "Discount Rate: {discount_rate}\n"
         "Bucket Size: {bucket_size}\n"
         "C Rate: {c_rate}\n"
@@ -435,8 +489,6 @@ def simulate_period_hourly_products(
     ).format(
         start_day=start_day,
         end_day=end_day,
-        threshold=threshold,
-        threshold_abs_min=threshold_abs_min,
         discount_rate=discount_rate,
         bucket_size=bucket_size,
         c_rate=c_rate,
@@ -447,91 +499,181 @@ def simulate_period_hourly_products(
 
     logger.info(log_message)
 
-    current_day = start_day
+    # Wie im Original: genau EIN Tag, current_day = start_day (auf 00:00 gesetzt)
+    current_day = start_day.replace(hour=0, minute=0, second=0, microsecond=0)
 
-    current_day = current_day.replace(hour=0, minute=0, second=0, microsecond=0)
-
+    # Trades-Sammel-DF (Day-Ahead + Intraday)
     all_trades = pd.DataFrame(
         columns=["execution_time", "side", "quantity", "price", "product", "profit"]
     )
 
-    if day_ahead_trades_drl is not None:
-        all_trades = pd.concat([all_trades, day_ahead_trades_drl])
+    # DRL-Day-Ahead-Trades für diesen Tag ableiten (falls vorhanden)
+    day_ahead_trades_drl = None
+    if drl_output is not None:
+        try:
+            day_ahead_trades_drl = derive_day_ahead_trades_from_drl_output(
+                drl_output, current_day
+            )
+            all_trades = pd.concat([all_trades, day_ahead_trades_drl], ignore_index=True)
+        except KeyError:
+            # Kein DRL-Output für diesen Tag → ohne DA-Startposition
+            day_ahead_trades_drl = None
 
+    # Zeitgrenzen (wie im Original)
     gate_closure_day_ahead = current_day - pd.Timedelta(days=1) + pd.Timedelta(hours=13)
     trading_start = current_day - pd.Timedelta(hours=8)
     trading_end = current_day + pd.Timedelta(days=1)
+
     execution_time_start = trading_start
     execution_time_end = trading_start + pd.Timedelta(minutes=bucket_size)
 
-    while execution_time_end < trading_end:
-        vwap = get_average_prices(
-            cursor,
-            "BUY",
-            execution_time_start,
-            execution_time_end,
-            trading_end,
-            min_trades=min_trades,
+    # VWAP-Matrix für diesen Tag laden
+    try:
+        vwap_matrix = load_vwap_matrix_for_day(current_day, base_path=precomputed_vwap_path)
+    except FileNotFoundError as e:
+        logger.warning(f"Keine vorcomputierten VWAPs für {current_day:%Y-%m-%d}: {e}")
+        # Rückgabe: leeres Reporting (wie create_quarterhourly_reporting bei leeren Trades)
+        empty_trades = pd.DataFrame(
+            columns=["execution_time", "side", "quantity", "price", "product", "profit"]
         )
+        return create_quarterhourly_reporting(empty_trades, start_day=current_day)
 
+    # (optional) Zähler, falls du noch Debug/Timing machen willst
+    db_time_total = 0.0
+    solver_time_total = 0.0
+    vwap_csv_time_total = 0.0
+
+    # Solange Intraday-Trading-Fenster offen ist
+    while execution_time_end < trading_end:
+        # VWAP für dieses Bucket-Ende aus der Matrix holen
+        t0 = time.perf_counter()
+        vwap = get_vwap_from_matrix_for_bucket(
+            vwap_matrix,
+            execution_time_end=execution_time_end,
+            end_date=trading_end,
+        )
+        db_time_total += time.perf_counter() - t0
+
+        #Optional: Logging der VWAPs in CSV
+        # ----------------------------------------------------
+        # vwaps_for_logging = (
+        #     vwap.copy().rename(columns={"price": execution_time_end}).T
+        # )
+        # year = start_day.year
+        # path = os.path.join(
+        #     "output",
+        #     "single_market",
+        #     "rolling_intrinsic",
+        #     "ri_basic",
+        #     "qh",
+        #     str(year),
+        #     "bs"
+        #     + str(bucket_size)
+        #     + "cr"
+        #     + str(c_rate)
+        #     + "rto"
+        #     + str(roundtrip_eff)
+        #     + "mc"
+        #     + str(max_cycles)
+        #     + "mt"
+        #     + str(min_trades),
+        # )
+        # vwappath = os.path.join(path, "vwap")
+        # os.makedirs(vwappath, exist_ok=True)
+        # vwap_filename = os.path.join(
+        #     vwappath, "vwaps_" + current_day.strftime("%Y-%m-%d") + ".csv"
+        # )
+        #
+        # t_csv0 = time.perf_counter()
+        # if not os.path.exists(vwap_filename):
+        #     vwaps_for_logging.to_csv(
+        #         vwap_filename,
+        #         mode="a",
+        #         header=True,
+        #         index=True,
+        #     )
+        # elif os.path.exists(vwap_filename) and (execution_time_start == trading_start):
+        #     os.remove(vwap_filename)
+        #     vwaps_for_logging.to_csv(
+        #         vwap_filename,
+        #         mode="a",
+        #         header=False,
+        #         index=True,
+        #     )
+        # else:
+        #     vwaps_for_logging.to_csv(
+        #         vwap_filename,
+        #         mode="a",
+        #         header=False,
+        #         index=True,
+        #     )
+        # vwap_csv_time_total += time.perf_counter() - t_csv0
+        # ----------------------------------------------------
+
+        # Bisherige Netto-Positionen inkl. DA-Trades
         net_trades = get_net_trades(all_trades, trading_end)
 
+        # Wenn für dieses Bucket gar keine VWAP-Preise vorhanden sind → weiter
         if vwap["price"].isnull().all():
             execution_time_start = execution_time_end
             execution_time_end = execution_time_start + pd.Timedelta(
                 minutes=bucket_size
             )
             continue
-        else:
-            try:
-                results, trades, profit = run_optimization_quarterhours_repositioning(
-                    vwap,
-                    execution_time_start,
-                    1,
-                    c_rate,
-                    roundtrip_eff,
-                    max_cycles,
-                    threshold,
-                    threshold_abs_min,
-                    discount_rate,
-                    net_trades,
-                )
-                all_trades = pd.concat([all_trades, trades])
-            except ValueError:  # TODO: see if ValueError is right
-                execution_time_start = execution_time_end
-                execution_time_end = execution_time_start + pd.Timedelta(
-                    minutes=bucket_size
-                )
 
-                continue
+        # Sonst Optimierung laufen lassen
+        try:
+            t_s0 = time.perf_counter()
+            results, trades, profit = run_optimization_quarterhours_repositioning(
+                vwap,
+                execution_time_start,
+                1,
+                c_rate,
+                roundtrip_eff,
+                max_cycles,
+                discount_rate,
+                net_trades,
+            )
+            solver_time_total += time.perf_counter() - t_s0
 
+            # neue Trades anhängen
+            all_trades = pd.concat([all_trades, trades], ignore_index=True)
+        except ValueError:
+            print("Error in optimization")
+            print("execution_time_start: ", execution_time_start)
+            execution_time_start = execution_time_end
+            execution_time_end = execution_time_start + pd.Timedelta(
+                minutes=bucket_size
+            )
+            continue
+
+        # nächstes Bucket
         execution_time_start = execution_time_end
         execution_time_end = execution_time_start + pd.Timedelta(minutes=bucket_size)
 
+    # Day-Ahead-Trades für Reporting entfernen
     if day_ahead_trades_drl is not None:
-        all_trades = all_trades[all_trades["execution_time"] != gate_closure_day_ahead]
+        all_trades_reporting = all_trades[
+            all_trades["execution_time"] != gate_closure_day_ahead
+        ]
+    else:
+        all_trades_reporting = all_trades
 
-    reporting = create_hourly_reporting(all_trades=all_trades, start_day=start_day)
-    current_log_dir = get_current_dir_version(LOGGING_PATH_COORDINATED)
-    logfile_path = os.path.join(
-        current_log_dir, COORDINATED_STACKED_RI_QH_TRAINING_OUTPUT_CSV
+    # Reporting-DF im gleichen Format wie simulate_period_quarterhourly_products
+    reporting = create_quarterhourly_reporting(
+        all_trades=all_trades_reporting, start_day=current_day
     )
-
-    # if not os.path.isfile(logfile_path):
-    #     reporting.to_csv(logfile_path, header=True)
-    # else:
-    #     reporting.to_csv(logfile_path, mode="a", header=False)
 
     return reporting
 
 
-def create_hourly_reporting(all_trades, start_day):
+def create_quarterhourly_reporting(all_trades, start_day):
     df = all_trades.copy()
     if df.empty:
         complete_index = pd.date_range(
             start_day,
-            start_day + pd.Timedelta(days=1) - pd.Timedelta(hours=1),
-            freq="h",
+            start_day + pd.Timedelta(days=1) - pd.Timedelta(minutes=15),
+            freq="15min",
         )
         complete_df = pd.DataFrame(
             index=complete_index, columns=["net_quantity", "vwap", "total_profit"]
@@ -543,10 +685,12 @@ def create_hourly_reporting(all_trades, start_day):
     df["net_quantity"] = df.apply(
         lambda x: -x["quantity"] if x["side"] == "buy" else x["quantity"], axis=1
     )
-    net_quantity = df.groupby("product")["net_quantity"].sum()
+    net_quantity = df.groupby("product")["net_quantity"].sum() / 4
 
-    df["vwap"] = df["price"] * df["quantity"]
-    vwap = df.groupby("product").apply(lambda x: x["vwap"].sum() / x["quantity"].sum())
+    df["vwap"] = df["price"] * df["quantity"] / 4
+    vwap = df.groupby("product").apply(
+        lambda x: x["vwap"].sum() / x["quantity"].sum() / 4
+    )
 
     total_profit = df.groupby("product")["profit"].sum()
     summary = pd.DataFrame(
@@ -555,8 +699,8 @@ def create_hourly_reporting(all_trades, start_day):
     summary = summary.reindex(
         index=pd.date_range(
             start_day,
-            start_day + pd.Timedelta(days=1) - pd.Timedelta(hours=1),
-            freq="h",
+            start_day + pd.Timedelta(days=1) - pd.Timedelta(minutes=15),
+            freq="15min",
         )
     )
     summary.index.name = "product"
