+++ src/shared/config.py
@@ -16,8 +16,8 @@ BUCKET_SIZE = 15
 MIN_TRADES = 10
 
 # Define model horizon
-START = pd.Timestamp(year=2019, month=1, day=1, tz="Europe/Berlin")
-END = pd.Timestamp(year=2024, month=1, day=1, tz="Europe/Berlin")
+START = pd.Timestamp(year=2019, month=10, day=25, tz="Europe/Berlin")
+END = pd.Timestamp(year=2019, month=11, day=30, tz="Europe/Berlin")
 
 # ----------------------------------------------
 # MODELLING CONFIGURATIONS
